+++ src/main/java/com/example/parking/repository/CarRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface CarRepository extends JpaRepository<Car, Integer> {
-    List<Car> findByUserUserId(Integer userId);
+public interface CarRepository extends JpaRepository<Car, Long> {
+    List<Car> findByUserUserId(Long userId);
 }
